chore: refactor code and update filters implementation
