+++ 01-javascript/empleos.html
@@ -26,7 +26,7 @@
             </nav>
         <div class="header-actions">
             <button>Subir CV</button>
-            <img src="avatar.svg" alt="User Avatar">
+            <dev-jobs-avatar size="32" service="github" username="botandres02"></dev-jobs-avatar>
         </div>
     </header>
     <main>
@@ -85,7 +85,7 @@
                 <h2>Resultados de búsqueda</h2>
             </div>
             <div class="jobs-listings">
-                <article class="job-listing-card" data-modalidad="remoto">
+                <!-- <article class="job-listing-card" data-modalidad="remoto">
                     <div>
                         <h3>Ingeniero de Software</h3>
                         <small>Tech Solutions Inc. | Remoto</small>
@@ -133,7 +133,7 @@
                             GCP.</p>
                     </div>
                     <button class="button-apply-job">Aplicar</button>
-                </article>
+                </article> -->
             </div>
 
             <nav class="pagination">
@@ -159,7 +159,7 @@
     <footer>
         <small>&copy; 2025 DevJobs. Todos los derechos reservados.</small>
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 
 </html>
