+++ 01-javascript/main.js
@@ -1,0 +1,4 @@
+import './scripts/apply-button.js';
+import './scripts/filters.js';
+import './scripts/fetch-data.js';
+import './components/dev-jobs-avatar.js';
